+++ demo/src/main/java/payroll/Manager/ManagerNotFoundException.java
@@ -2,7 +2,7 @@ package payroll.Manager;
 
 public class ManagerNotFoundException extends RuntimeException {
 
-	ManagerNotFoundException(Long id) {
+	ManagerNotFoundException(int id) {
         super("Could not find Manager " + id);
     }
 }
